src/python/pubsub-client-pi/requirements.txt
google-api-python-client==1.6.4
google-auth-httplib2==0.0.3
google-auth==1.2.1
google-cloud-pubsub==0.29.4
cryptography==2.1.4
pyjwt==1.5.3
paho-mqtt==1.3.1
gpiozero==1.4.0
